speech/requirements.txt
openai-whisper
openai
pytubefix

# app
gradio